Tests fix
Snapshots updated with fingerprints

Fix snapshots

Fixed default ruleset handling

Fix tests

Fixed tests

Fixed ASM_Data configs processor

Fixed snapshot version

Fixed one click issue

Update waf version in smoke tests snapshots

Fixed some log messages

+++ tracer/src/Datadog.Trace/AppSec/Waf/ReturnTypes.Managed/DiagnosticResultUtils.cs
@@ -139,7 +139,7 @@ internal struct WafStats
     {
     }
 
-    public uint TotalCount => (uint)(FailedCount + LoadedCount + SkippedCount);
+    public int TotalCount => (FailedCount + LoadedCount + SkippedCount);
 }
 
 #pragma warning restore SA1402 // File may only contain a single type
